Ajustes chicos 1 - Agregada la función de volver en agregar clase - Ajustado el encabezado en el menú principal y en importar clase - Nueva página para importar alumnos - Cambio en botones

+++ Teechers App/ClassTabs/Attendance/Cells/tchAttendanceColumnVC.m
@@ -42,7 +42,7 @@
         labelText = [statusArray objectAtIndex:[record.status integerValue]];
         
         // it the record is absent
-        if ([record.status isEqual: [NSString stringWithFormat:@"%i", tchAttendanceAbsent]]
+        if ([record.status isEqual: [NSString stringWithFormat:@"%li", (long)tchAttendanceAbsent]]
             && ([record.excused intValue] == tchAttendanceExcusedNo)) {
             
             // set the label color to red
@@ -61,7 +61,7 @@
 {
     
     // pass index to label
-    self.dataLabel.text = [NSString stringWithFormat:@"%i", currentIndex];
+    self.dataLabel.text = [NSString stringWithFormat:@"%li", (long)currentIndex];
     
 }
